src/backend/requirements.txt
fastapi
uvicorn
pydantic
requests
python-dotenv
sqlalchemy
psycopg2-binary
twilio
openai
